+++ hn/services/firebase/types.go
@@ -2,12 +2,14 @@ package firebase
 
 type hnItem struct {
 	ID          int    `json:"id"`
+	Type        string `json:"type"`
 	By          string `json:"by"`
 	Time        int64  `json:"time"`
 	Text        string `json:"text"`
 	Title       string `json:"title"`
 	URL         string `json:"url"`
 	Score       int    `json:"score"`
+	Parent      int    `json:"parent"`
 	Kids        []int  `json:"kids"`
 	Descendants int    `json:"descendants"`
 	Deleted     bool   `json:"deleted"`